requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Limiter==3.7.0
Flask-Caching==2.3.0
Werkzeug==3.0.3
gunicorn==22.0.0
